LPS-117060 Initialize contexts inside App component
