+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/index.js
@@ -16,32 +16,8 @@ import React from 'react';
 import {DndProvider} from 'react-dnd';
 import {HTML5Backend} from 'react-dnd-html5-backend';
 
-import {INIT} from '../app/actions/types';
 import App from './components/App';
-import {ControlsProvider} from './components/Controls';
 import {initializeConfig} from './config/index';
-import {reducer} from './reducers/index';
-import {StoreContextProvider} from './store/index';
-
-/**
- * Container component that sets up context that is global to the entire app.
- *
- * This is a separate functional component instead of being directly inlined in
- * this module's default export because hooks can only be used inside functional
- * components (the default export is not a functional component but rather a
- * function that returns a component).
- */
-function Container({state}) {
-	const initialState = reducer(state, {type: INIT});
-
-	return (
-		<StoreContextProvider initialState={initialState} reducer={reducer}>
-			<ControlsProvider>
-				<App />
-			</ControlsProvider>
-		</StoreContextProvider>
-	);
-}
 
 /**
  * Default application export.
@@ -54,7 +30,7 @@ export default function (data) {
 
 	return (
 		<DndProvider backend={HTML5Backend}>
-			<Container state={data.state} />
+			<App state={data.state} />
 		</DndProvider>
 	);
 }
